==> 0_GenerateDS.ipynb <==
# Generate Songs Dataset (random sample)

import pandas as pd


df = pd.read_excel('data/GenerateDataset.xlsx', sheet_name='Sheet1')

df = df[['track', 'artist', 'seq']]

df


df[df['artist']=="Pink Floyd"]

df = df.sample(n = 15000)

import pickle

with open('generatedDS.pkl', 'wb') as file:
    pickle.dump(df, file)
