==> 1_emotionAnalysisBERT.ipynb <==
# Song Lyrics -> Cleaning and Emotion Analysis with BERT

# pip install --upgrade pandas

import os
import pandas as pd
import string
import tensorflow as tf
import re
import ast
import pickle
import nltk
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize

with open('generatedDS.pkl', 'rb') as file:
    data_lyrics = pd.read_pickle(file)


data_lyrics = data_lyrics.dropna()

data_lyrics

data_lyrics = data_lyrics[data_lyrics['seq'].str.strip() != '']
data_lyrics

data_lyrics['clean_text'] = data_lyrics['seq'].str.lower().str.translate(str.maketrans('', '', string.punctuation))
data_lyrics

data_lyrics['clean_text'] = data_lyrics['clean_text'].apply(lambda x: re.sub(r"\n", " ", x))
data_lyrics['clean_text'] = data_lyrics['clean_text'].apply(lambda x: re.sub(r"x[0-9]+d", " ", x))
# data_lyrics['clean_text'] = data_lyrics['clean_text'].apply(lambda x: re.sub(r"_x[0-9]+D_\n", " ", x))
data_lyrics['clean_text'] = data_lyrics['clean_text'].apply(lambda x: re.sub(r"x[0-9]+", " ", x))
data_lyrics['clean_text'] = data_lyrics['clean_text'].apply(lambda x: re.sub(r"\s+", " ", x))
data_lyrics

#Take only first five hundred characters of each song to reduce text size
data_lyrics['clean_text'] = data_lyrics['clean_text'].str.slice(0, 500)
data_lyrics

# nltk.download('punkt')

column_name = 'clean_text'

nltk.download('stopwords')

stop_words = set(stopwords.words('english'))

def remove_stopwords(text):
    tokens = word_tokenize(text)
    filtered_tokens = [word for word in tokens if word.lower() not in stop_words]
    return ' '.join(filtered_tokens)

data_lyrics[column_name] = data_lyrics[column_name].apply(remove_stopwords)


data_lyrics

# with open('cleaned_generatedDS.pkl', 'wb') as file:
#     pickle.dump(data_lyrics, file)

# Obtain most important words on each lyric

# pip install -U scikit-learn



from sklearn.feature_extraction.text import TfidfVectorizer


# Create TF-IDF vectors
vectorizer = TfidfVectorizer()
tfidf_matrix = vectorizer.fit_transform(data_lyrics['clean_text'])

n_top_words = 10

top_words_list = []

for i in range(len(data_lyrics)):
    lyric_tfidf_scores = tfidf_matrix[i].toarray()[0]
    top_words_indices = lyric_tfidf_scores.argsort()[-n_top_words:][::-1]
    top_words = [vectorizer.get_feature_names_out()[index] for index in top_words_indices]
    top_words_list.append(top_words)
    # print(f"Top {n_top_words} words for song {i+1}: {top_words}")

data_lyrics['top_words'] = top_words_list

data_lyrics

# data_lyrics.to_excel("tfidfdatalyrics.xlsx")

data_lyrics = pd.read_excel("tfidfdatalyrics.xlsx")
data_lyrics

with open('tfidfdatalyrics.pkl', 'wb') as file:
    pickle.dump(data_lyrics, file)

# Obtain emotion for each song using BERT (Zero-Shot)

# !pip install transformers

import torch
import pandas as pd
import numpy as np
from transformers import AutoTokenizer, AutoModelForSequenceClassification, Trainer

# Create class for data preparation
class SimpleDataset:
    def __init__(self, tokenized_texts):
        self.tokenized_texts = tokenized_texts
    
    def __len__(self):
        return len(self.tokenized_texts["input_ids"])
    
    def __getitem__(self, idx):
        return {k: v[idx] for k, v in self.tokenized_texts.items()}

# pip install --upgrade accelerate

# load tokenizer and model, create trainer
model_name = "j-hartmann/emotion-english-distilroberta-base"
tokenizer = AutoTokenizer.from_pretrained(model_name)
model = AutoModelForSequenceClassification.from_pretrained(model_name)
trainer = Trainer(model=model)

pred_texts = data_lyrics['clean_text'].tolist()

tokenized_texts = tokenizer(pred_texts,truncation=True,padding=True)
pred_dataset = SimpleDataset(tokenized_texts)

predictions = trainer.predict(pred_dataset)

# Transform predictions to labels
preds = predictions.predictions.argmax(-1)
labels = pd.Series(preds).map(model.config.id2label)
scores = (np.exp(predictions[0])/np.exp(predictions[0]).sum(-1,keepdims=True)).max(1)

# scores raw
temp = (np.exp(predictions[0])/np.exp(predictions[0]).sum(-1,keepdims=True))

# work in progress
# container
anger = []
disgust = []
fear = []
joy = []
neutral = []
sadness = []
surprise = []


#    'Amusement': 0,
#     'Anger': 1,
#     'Awe': 2,
#     'Contentment': 3,
#     'Excitement': 4,
#     'Fear': 5,
#     'Sadness': 6,
#     'Something else': 7



# extract scores (as many entries as exist in pred_texts)
for i in range(len(pred_texts)):
  anger.append(temp[i][0])
  disgust.append(temp[i][1])
  fear.append(temp[i][2])
  joy.append(temp[i][3])
  neutral.append(temp[i][4])
  sadness.append(temp[i][5])
  surprise.append(temp[i][6])

# Create DataFrame with texts, predictions, labels, and scores
df = pd.DataFrame(list(zip(pred_texts,preds,labels,scores,  anger, disgust, fear, joy, neutral, sadness, surprise)), columns=['text','pred','label','score', 'anger', 'disgust', 'fear', 'joy', 'neutral', 'sadness', 'surprise'])
# df.to_excel("dataframe_sentimientos.xlsx")
df

df

# SAVE TO PKL

with open('dataframe_sentimientos.pkl', 'wb') as file:
    pickle.dump(df, file)
