==> 2_GenerateImageDSs.ipynb <==
# Song TF-IDF Embeddings and ArtEmis Descriptions Embeddings

import pickle
import pandas as pd
import numpy as np
import ast
import torch

with open('dataframe_sentimientos.pkl', 'rb') as file:
    df_sentimientos = pickle.load(file)

with open('tfidfdatalyrics.pkl', 'rb') as file:
    df_tfidfdatalyrics = pickle.load(file)

df_sentimientos = df_sentimientos.drop(columns = ["score", "anger", "disgust", "fear", "joy", "neutral", "sadness", "surprise", "text"])
df_sentimientos

df_tfidfdatalyrics = df_tfidfdatalyrics.drop(columns=["Unnamed: 0", "seq"])
df_tfidfdatalyrics

df_merged = pd.concat([df_tfidfdatalyrics, df_sentimientos], axis=1)
df_merged

df_merged = df_merged[df_merged['label'] != 'disgust']
df_merged

import ast
df_merged['top_words_sentence'] = df_merged['top_words'].apply(lambda x: ' '.join(ast.literal_eval(x)))
df_merged

emotion_mapping = {
    'sadness': 'sadness',
    'anger': 'anger',
    'surprise': 'excitement',
    'joy': 'contentment',
    'neutral': 'something else',
    'fear': 'fear'
}

df_merged['label'] = df_merged['label'].map(emotion_mapping)
df_merged

excel_file_path = 'dataframe_tfidf_emotion.xlsx'

df_merged.to_excel(excel_file_path, index=False)

emotion_dfs = {}
for emotion in df_merged['label'].unique():
    emotion_dfs[emotion] = df_merged[df_merged['label'] == emotion].copy()

for emotion, emotion_df in emotion_dfs.items():
    print(f"Dataframe for {emotion}:")
    print(len(emotion_df))
    print()

df_artemis = pd.read_csv('data/artemis_dataset_release_v0.csv')
df_artemis

emotion_dfs_artemis = {}
for emotion in df_artemis['emotion'].unique():
    emotion_dfs_artemis[emotion] = df_artemis[df_artemis['emotion'] == emotion].copy()

for emotion, emotion_df_artemis in emotion_dfs_artemis.items():
    print(f"Dataframe for {emotion}:")
    #print(emotion_df_artemis.head(1))
    print(len(emotion_df_artemis))
    print()

# pip install -U sentence-transformers

### GENERATE EMBEDDINGS

from sentence_transformers import SentenceTransformer

model = SentenceTransformer('sentence-transformers/all-MiniLM-L6-v2')

# This is a sentence-transformers model: It maps sentences & paragraphs to 
#a 384 dimensional dense vector space and can be used for tasks like clustering or semantic search.

# print( emotion_dfs_artemis)

embeddings_lyrics = []
for emotion, emotion_df in emotion_dfs.items():
    # print(emotion, emotion_df)
    # print(emotion_df['top_words_sentence'])
    embeddings_lyrics.append(model.encode(emotion_df['top_words_sentence'].tolist(), convert_to_tensor=True))

embeddings_artemis = []
for emotion, emotion_df_artemis in emotion_dfs_artemis.items():
    embeddings_artemis.append(model.encode(emotion_df_artemis['utterance'].tolist(), convert_to_tensor=True))




# with open("embeddings_lyrics.pkl", "wb") as embeddings_lyrics_file:
#     pickle.dump(embeddings_lyrics, embeddings_lyrics_file)

# 
# with open("embeddings_artemis.pkl", "wb") as embeddings_artemis_file:
    # pickle.dump(embeddings_artemis, embeddings_artemis_file)

### Remember to save the embeddings... Uncomment previous lines if you did not
