==> 3_pickSongGenerateDS.ipynb <==
# Let a user choose an artist and song based on which images are to be selected and GAN images generated

import pickle
import pandas as pd
import numpy as np
import ast
import torch

with open('embeddings_lyrics.pkl', 'rb') as file:
    embeddings_lyrics = pickle.load(file)

with open('embeddings_artemis.pkl', 'rb') as file:
    embeddings_artemis = pickle.load(file)

path = "dataframe_tfidf_emotion.xlsx"
dataframe_songs = pd.read_excel(path)


### Select a song from the dataframe

##### Feel free to filter by artist:

artist = input("Enter artist name: ")

dataframe_songs[dataframe_songs['artist'] == artist]

##### Select the song index:

song_index = int(input("Enter the song index: "))

song = dataframe_songs.loc[song_index]
print("You have chosen: {}".format(song['track']), "by {}".format(song['artist']))

song_emotion = dataframe_songs.loc[song_index]['label']
song_emotion

emotion_mapping = {
    'contentment': 0,
    'sadness': 1,
    'excitement': 2,
    'fear': 3,
    'anger': 4,
    'something else': 5
}

song_emotion_index = emotion_mapping[song_emotion]
song_emotion_index

# Find the correct index on the dataframe

emotion_dfs = {}
for emotion in dataframe_songs['label'].unique():
    emotion_dfs[emotion] = dataframe_songs[dataframe_songs['label'] == emotion].copy()

emotion_df = emotion_dfs[song_emotion]
matching_rows = emotion_df[(emotion_df['track'] == song['track']) & (emotion_df['artist'] == song['artist'])].index

for row in matching_rows:
    print(row)

i = 0
for index, row in emotion_df.iterrows():
    if row['track'] == song['track'] and row['artist'] == song['artist']:
        print("Row number:", index, i)
        index = i
        break
    i+=1

df_artemis = pd.read_csv('data/artemis_dataset_release_v0.csv')

emotion_dfs_artemis = {}
for emotion in df_artemis['emotion'].unique():
    emotion_dfs_artemis[emotion] = df_artemis[df_artemis['emotion'] == emotion].copy()

print(embeddings_lyrics[0].shape)
print(embeddings_lyrics[1].shape)
print(embeddings_lyrics[2].shape)
print(embeddings_lyrics[3].shape)
print(embeddings_lyrics[4].shape)
print(embeddings_lyrics[5].shape)

# embeddings_lyrics[1][4689]

print(index)
if song_emotion_index == 0:
    similarity = torch.matmul(embeddings_lyrics[0][index], embeddings_artemis[2].T)
elif song_emotion_index == 1:
    similarity = torch.matmul(embeddings_lyrics[1][index], embeddings_artemis[1].T)
elif song_emotion_index == 2:
    similarity = torch.matmul(embeddings_lyrics[2][index], embeddings_artemis[5].T)
elif song_emotion_index == 3:
    similarity = torch.matmul(embeddings_lyrics[3][index], embeddings_artemis[6].T)
elif song_emotion_index == 4:
    similarity = torch.matmul(embeddings_lyrics[4][index], embeddings_artemis[8].T)
elif song_emotion_index == 5:
    similarity = torch.matmul(embeddings_lyrics[5][index], embeddings_artemis[0].T)

print(similarity)
print(similarity.shape)


top_indices = np.argsort(similarity)[-1500:]
#print(top_indices.numpy().tolist())

print(emotion_dfs[song_emotion].iloc[index]['top_words'])
most_similar_sentences = emotion_dfs_artemis[song_emotion].iloc[top_indices.numpy().tolist()]
most_similar_sentences

# emotion_dfs_artemis[song_emotion].iloc[26833]

## Create dataset for song

import os
import shutil

path_destination = 'pickSongGenerateDS/no_label/'
for index, row in most_similar_sentences.iterrows():
    directory = 'data/' + row['art_style']
    file = row['painting'] + '.jpg'
    
    for root, dirs, files in os.walk(directory):
        if file in files:
            source_file_path = os.path.join(root, file)
            destination_file_path = os.path.join(path_destination, file)
            shutil.copyfile(source_file_path, destination_file_path)
            #print("File copied successfully!")
    
    

#### The generated images dataset is located in pickSongGeneratedDS/
