==> 5_GenerateImageVideo.ipynb <==
# Generate Video Based On GAN Generated Images

# pip install moviepy

# !pip install opencv-python

import cv2
import os

input_dir = 'generated_images_GAN'

output_file = 'generated_video/video.avi'

frame_rate = 68

image_files = os.listdir(input_dir)

first_image = cv2.imread(os.path.join(input_dir, image_files[0]))
if first_image is None:
    raise ValueError("Unable to load the first image")

height, width, _ = first_image.shape

fourcc = cv2.VideoWriter_fourcc(*'XVID')
video_writer = cv2.VideoWriter(output_file, fourcc, frame_rate, (width, height))

for i in range(len(image_files)-1):
    # Read the current and next image
    image1 = cv2.imread(os.path.join(input_dir, image_files[i]))
    image2 = cv2.imread(os.path.join(input_dir, image_files[i+1]))

    if image1 is None or image2 is None:
        print(f"Unable to load image at index {i}")
        continue

    # Resize the images to a larger size for smoother interpolation
    scale = 10 #
    resized_image1 = cv2.resize(image1, (width*scale, height*scale))
    resized_image2 = cv2.resize(image2, (width*scale, height*scale))

    
    alpha = 0.9 
    interpolated_image = cv2.addWeighted(resized_image1, 1 - alpha, resized_image2, alpha, 0)

    # Resize the interpolated image back to the original size
    interpolated_image = cv2.resize(interpolated_image, (width, height))

    # Write the interpolated frame to the video multiple times for smoother transition
    num_frames = 68 # Number of frames per interpolation (adjust as needed)
    for _ in range(num_frames):
        video_writer.write(interpolated_image)

video_writer.release()
